Create logsearcher.sh
The requirements and instructions can be found in the attached file in the same folder.

My enhanced functionality is that the code can automatically search all .csv files in the directory and let the search run on all of them simultaneously or any one at a time. This allows for the code to work in real life scenarios where someone might have to search through all the log files at once rather than just one or five of them at a given time. This also greatly improves the efficiency of the user and lets them save a lot of time.